% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Atlantic}
\alias{Atlantic}
\title{Atlantic core-top foram assemblages}
\format{
A data frame with 1093 rows and 33 variables.
summ50 is 50m water temperature of the warmest season
}
\source{
\itemize{
\item Foraminifera data \doi{10.1594/PANGAEA.227322}
\item Sea surface temperature data from WOA13
\url{https://www.ncei.noaa.gov/data/oceans/woa/WOA13/DATAv2/temperature/csv/decav/1.00/}
file 'woa13_decav_t13mn01v2.csv.gz'
}
}
\usage{
data(Atlantic)
}
\description{
A dataset containing over 1000 foram assemblages from the Atlantic from
Kucera et al (2005) and the 50m SST for the warmest season. Rare taxa and
co-located assemblages are removed.
}
\keyword{datasets}
