\name{palaeoSig-package}
\alias{palaeoSig-package}
\docType{package}
\title{
Significance Tests of Quantitative Palaeoenvironmental Reconstructions 
}
\description{
Significance tests for quantitative palaeoenvironmental reconstructions derived from transfer functions.


}
\details{
randomTF tests if the reconstruction explains more of the variance in a constrained ordination than a reconstruction derived from a transfer function trained on random environmental data
obscor.sig tests if the weighted correlation between the weighed average species optima and the species scores from a ordination constrained by the weighted averaging reconstruction is larger than that expected for random environmental data.
}                                      
\author{
Richard Telford \email{richard.telford@bio.uib.no}
}
\references{
Telford, R.J. and Birks, H.J.B. (In Revision) A novel method for assessing the statistical significance of quantitative reconstructions inferred from biotic assemblages. Quaternary Science Reviews. 
}

\keyword{ package }
\seealso{
\code{\link[rioja:rioja-package]{rioja}}, \code{\link{rda}}
}

